t: move testutils.go to cmd sub-package
