+++ lfs/scanner_git_test.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	. "github.com/git-lfs/git-lfs/lfs"
-	"github.com/git-lfs/git-lfs/test"
+	test "github.com/git-lfs/git-lfs/t/cmd/util"
 	"github.com/stretchr/testify/assert"
 )
