+++ t/cmd/lfstest-count-tests.go
@@ -30,7 +30,8 @@ func main() {
 		os.Exit(1)
 	}
 
-	ctx, _ := context.WithTimeout(context.Background(), 5*time.Second)
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
 
 	if err := acquire(ctx); err != nil {
 		fatal(err)
